api: split node_id.go
